Complete rework of the GEM bindings. Changed all 'int' in system interfaces to the real type 'short'. Prefixed some macros in gem.h due to collisions with other libs (object states -> OS_, object flags -> OF_, object colors -> OC_).
Programmers attention: you must adjust your source code to the new
bindings and macros names (for compatibility the old macro names
are available if __GEMLIB_OLDNAMES is defined before gem.h is
included.

+++ gemlib/a_appl_bvset.c
@@ -2,16 +2,19 @@
 #include "gem.h"
 
 
-int
-appl_bvset (int bvdisk, int bvhard)
+short
+appl_bvset (short bvdisk, short bvhard)
 {
 	aes_intin[0] = bvdisk;
 	aes_intin[1] = bvhard;
+	
 	aes_control[0] = 16;
 	aes_control[1] = 2;
 	aes_control[2] = 1;
 	aes_control[3] = 0;
 	aes_control[4] = 0;
+	
 	aes (&aes_params);
+	
 	return aes_intout[0];
 }
