+++ gemlib/a_appl_write.c
@@ -2,17 +2,21 @@
 #include "gem.h"
 
 
-int
-appl_write (int ApId, int Length, void *ApPbuff)
+short
+appl_write (short ApId, short Length, void *ApPbuff)
 {
 	aes_intin[0] = ApId;
 	aes_intin[1] = Length;
+	
 	aes_addrin[0] = (long) ApPbuff;
+	
 	aes_control[0] = 12;
 	aes_control[1] = 2;
 	aes_control[2] = 1;
 	aes_control[3] = 1;
 	aes_control[4] = 0;
+	
 	aes (&aes_params);
+	
 	return aes_intout[0];
 }
