+++ gemlib/a_menu_icheck.c
@@ -2,17 +2,21 @@
 #include "gem.h"
 
 
-int
-menu_icheck (void *Tree, int Item, int CheckFlag)
+short
+menu_icheck (void *Tree, short Item, short CheckFlag)
 {
 	aes_intin[0] = Item;
 	aes_intin[1] = CheckFlag;
+	
 	aes_addrin[0] = (long) Tree;
+	
 	aes_control[0] = 31;
 	aes_control[1] = 2;
 	aes_control[2] = 1;
 	aes_control[3] = 1;
 	aes_control[4] = 0;
+	
 	aes (&aes_params);
+	
 	return aes_intout[0];
 }
