+++ gemlib/a_wind_set_grect.c
@@ -3,7 +3,7 @@
 
 
 void
-wind_set_grect (int WindowHandle, int What, GRECT * r)
+wind_set_grect (short WindowHandle, short What, GRECT * r)
 {
 	aes_intin[0] = WindowHandle;
 	aes_intin[1] = What;
@@ -11,10 +11,12 @@ wind_set_grect (int WindowHandle, int What, GRECT * r)
 	aes_intin[3] = r->g_y;
 	aes_intin[4] = r->g_w;
 	aes_intin[5] = r->g_h;
+	
 	aes_control[0] = 105;
 	aes_control[1] = 6;
 	aes_control[2] = 1;
 	aes_control[3] = 0;
 	aes_control[4] = 0;
+	
 	aes (&aes_params);
 }
