+++ gemlib/n_v_get_outline.c
@@ -2,9 +2,9 @@
 #include "gemx.h"
 
 
-int
-v_get_outline (int handle, int index, int x_offset, int y_offset, short *pts,
-	       char *flags, int max_pts)
+short
+v_get_outline (short handle, short index, short x_offset, short y_offset, short *pts,
+	       char *flags, short max_pts)
 {
 	vdi_intin[0] = index;
 	vdi_intin[1] = max_pts;
