+++ gemlib/v_eeol.c
@@ -3,7 +3,7 @@
 
 
 void
-v_eeol (int handle)
+v_eeol (short handle)
 {
 	vdi_control[0] = 5;
 	vdi_control[1] = 0;
